+++ config/jest-config-ibm-cloud-cognitive/index.js
@@ -9,7 +9,40 @@
 
 module.exports = {
   collectCoverageFrom: ['src/**/*.js', '!**/*.stories.js'],
-  coverageReporters: [['html', 'text', { skipEmpty: true }]],
+  coverageReporters: [
+    [
+      'html',
+      {
+        skipEmpty: true,
+        // this next part doesn't actually seem to work
+        // see https://github.com/facebook/jest/issues/9734
+        watermark: {
+          statements: [80, 100],
+          lines: [80, 100],
+          functions: [80, 100],
+          branches: [80, 100],
+        },
+      },
+    ],
+  ],
+  // set the global coverage threshold, because that supplies the default for
+  // the upper watermark, then nullify the global coverage threshold so it
+  // doesn't actually cause jest to fail whenever anything falls short of 100%!
+  // This can be removed if/when there's a way to set reporter watermarks directly.
+  coverageThreshold: {
+    global: {
+      branches: 100,
+      functions: 100,
+      lines: 100,
+      statements: 100,
+    },
+    '**/*.js': {
+      branches: 0,
+      functions: 0,
+      lines: 0,
+      statements: 0,
+    },
+  },
   moduleFileExtensions: ['js', 'json', 'node'],
   reporters: ['default'],
   setupFiles: [require.resolve('./setup/setupFiles')],